No need for size_t in ngram_prob
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@12662 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/sphinxbase/ngram_model.h
@@ -304,7 +304,7 @@ int32 ngram_probv(ngram_model_t *model, const char *word, ...);
  * unigram weight (interpolation with uniform) is not removed.
  */
 SPHINXBASE_EXPORT
-int32 ngram_prob(ngram_model_t *model, const char *const *words, size_t n);
+int32 ngram_prob(ngram_model_t *model, const char *const *words, int32 n);
 
 /**
  * Quick "raw" probability lookup for a general N-Gram.
